requirements.txt
streamlit>=1.35.0
google-genai>=1.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
langchain>=0.2.0
langchain-community>=0.2.0
faiss-cpu>=1.8.0
langchain-text-splitters
langchain-core
 